requirements.txt
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth-oauthlib==0.4.6
google-auth==2.6.0
google-cloud-bigquery==2.34.0
pandas==1.4.1
oauth2client==4.1.3
spotipy==2.19.0
pyarrow==7.0.0